// packages/ai/src/providers/devin/proto/exa/auth_pb/auth_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts"
// @generated from file exa/auth_pb/auth.proto (package exa.auth_pb, syntax proto3)
/* eslint-disable */

import type { Message } from "@bufbuild/protobuf";
import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
import { file_buf_validate_validate } from "../../buf/validate/validate_pb.js";
import type { Metadata } from "../codeium_common_pb/codeium_common_pb.js";
import { file_exa_codeium_common_pb_codeium_common } from "../codeium_common_pb/codeium_common_pb.js";

/**
 * Describes the file exa/auth_pb/auth.proto.
 */
export const file_exa_auth_pb_auth: GenFile =
	/*@__PURE__*/
	fileDesc(
		"ChZleGEvYXV0aF9wYi9hdXRoLnByb3RvEgtleGEuYXV0aF9wYiJGChFHZXRVc2VySnd0UmVxdWVzdBIxCghtZXRhZGF0YRgBIAEoCzIfLmV4YS5jb2RlaXVtX2NvbW1vbl9wYi5NZXRhZGF0YSJFChJHZXRVc2VySnd0UmVzcG9uc2USEAoIdXNlcl9qd3QYASABKAkSHQoVY3VzdG9tX2FwaV9zZXJ2ZXJfdXJsGAIgASgJMlwKC0F1dGhTZXJ2aWNlEk0KCkdldFVzZXJKd3QSHi5leGEuYXV0aF9wYi5HZXRVc2VySnd0UmVxdWVzdBofLmV4YS5hdXRoX3BiLkdldFVzZXJKd3RSZXNwb25zZWIGcHJvdG8z",
		[file_buf_validate_validate, file_exa_codeium_common_pb_codeium_common],
	);

/**
 * @generated from message exa.auth_pb.GetUserJwtRequest
 */
export type GetUserJwtRequest = Message<"exa.auth_pb.GetUserJwtRequest"> & {
	/**
	 * @generated from field: exa.codeium_common_pb.Metadata metadata = 1;
	 */
	metadata?: Metadata | undefined;
};

/**
 * Describes the message exa.auth_pb.GetUserJwtRequest.
 * Use `create(GetUserJwtRequestSchema)` to create a new message.
 */
export const GetUserJwtRequestSchema: GenMessage<GetUserJwtRequest> =
	/*@__PURE__*/
	messageDesc(file_exa_auth_pb_auth, 0);

/**
 * @generated from message exa.auth_pb.GetUserJwtResponse
 */
export type GetUserJwtResponse = Message<"exa.auth_pb.GetUserJwtResponse"> & {
	/**
	 * @generated from field: string user_jwt = 1;
	 */
	userJwt: string;

	/**
	 * @generated from field: string custom_api_server_url = 2;
	 */
	customApiServerUrl: string;
};

/**
 * Describes the message exa.auth_pb.GetUserJwtResponse.
 * Use `create(GetUserJwtResponseSchema)` to create a new message.
 */
export const GetUserJwtResponseSchema: GenMessage<GetUserJwtResponse> =
	/*@__PURE__*/
	messageDesc(file_exa_auth_pb_auth, 1);

/**
 * @generated from service exa.auth_pb.AuthService
 */
export const AuthService: GenService<{
	/**
	 * @generated from rpc exa.auth_pb.AuthService.GetUserJwt
	 */
	getUserJwt: {
		methodKind: "unary";
		input: typeof GetUserJwtRequestSchema;
		output: typeof GetUserJwtResponseSchema;
	};
}> = /*@__PURE__*/ serviceDesc(file_exa_auth_pb_auth, 0);